Update code
Fix some wording and references to dataframe columns. Enforced consistent styling of column names

By making the switch to `NORMALIZED TOTAL COST`, there are now no longer any anomalous values for `EVENT SUBGROUP`. I've left the code in for now, consider finding another field that has weird data that needs to be cleaned. It's an important part of the process that should be shown.

I had left a few comments throughout the notebook (headers starting with #) in sections that you could expand more on the process, make sure to check them out and remove them once you'd added more to it.